+++ mex/backend/graph/exceptions.py
@@ -1,3 +1,6 @@
+from pydantic import ValidationError
+from pydantic_core import ErrorDetails
+
 from mex.common.exceptions import MExError
 
 
@@ -11,3 +14,9 @@ class MultipleResultsFoundError(MExError):
 
 class InconsistentGraphError(MExError):
     """Exception raised for inconsistencies found in the graph database."""
+
+    def errors(self) -> list[ErrorDetails]:
+        """Details about each underlying error."""
+        if isinstance(self.__cause__, ValidationError):
+            return self.__cause__.errors()
+        return []