+++ mex/backend/main.py
@@ -10,8 +10,9 @@ from fastapi.openapi.utils import get_openapi
 from pydantic import BaseModel, ValidationError
 
 from mex.backend.auxiliary.wikidata import router as wikidata_router
-from mex.backend.exceptions import handle_uncaught_exception, handle_validation_error
+from mex.backend.exceptions import handle_detailed_error, handle_uncaught_exception
 from mex.backend.extracted.main import router as extracted_router
+from mex.backend.graph.exceptions import InconsistentGraphError
 from mex.backend.identity.main import router as identity_router
 from mex.backend.ingest.main import router as ingest_router
 from mex.backend.logging import UVICORN_LOGGING_CONFIG
@@ -49,7 +50,7 @@ def create_openapi_schema() -> dict[str, Any]:
         summary=app.summary,
         description=app.description,
         routes=app.routes,
-        servers=[dict(url=settings.backend_api_url)],
+        servers=[{"url": settings.backend_api_url}],
     )
     for identifier in chain(EXTRACTED_IDENTIFIER_CLASSES, MERGED_IDENTIFIER_CLASSES):
         name = identifier.__name__
@@ -78,11 +79,11 @@ app = FastAPI(
         "The MEx API includes endpoints for multiple use-cases, "
         "e.g. for extractor pipelines, the MEx editor or inter-departmental access."
     ),
-    contact=dict(
-        name="RKI MEx Team",
-        email="mex@rki.de",
-        url="https://github.com/robert-koch-institut/mex-backend",
-    ),
+    contact={
+        "name": "RKI MEx Team",
+        "email": "mex@rki.de",
+        "url": "https://github.com/robert-koch-institut/mex-backend",
+    },
     lifespan=lifespan,
     version="v0",
 )
@@ -110,7 +111,8 @@ def check_system_status() -> SystemStatus:
 
 
 app.include_router(router)
-app.add_exception_handler(ValidationError, handle_validation_error)
+app.add_exception_handler(InconsistentGraphError, handle_detailed_error)
+app.add_exception_handler(ValidationError, handle_detailed_error)
 app.add_exception_handler(Exception, handle_uncaught_exception)
 app.add_middleware(
     CORSMiddleware,