+++ tests/graph/test_query.py
@@ -4,7 +4,7 @@ from pydantic import ValidationError
 from mex.backend.graph.query import QueryBuilder, render_constraints
 
 
-@pytest.fixture()
+@pytest.fixture
 def query_builder() -> QueryBuilder:
     builder = QueryBuilder.get()
     builder._env.globals.update(
@@ -74,7 +74,7 @@ YIELD currentStatus;"""
             True,
             True,
             """\
-CALL {
+CALL () {
     CALL db.index.fulltext.queryNodes("search_index", $query_string)
     YIELD node AS hit, score
     CALL {
@@ -195,7 +195,7 @@ def test_fetch_extracted_items(
             True,
             True,
             """\
-CALL {
+CALL () {
     CALL db.index.fulltext.queryNodes("search_index", $query_string)
     YIELD node AS hit, score
     OPTIONAL MATCH (n:AdditiveThis|AdditiveThat|AdditiveOther|ExtractedThis|ExtractedThat|ExtractedOther)-[:stableTargetId]->(merged:MergedThis|MergedThat|MergedOther)
@@ -206,7 +206,7 @@ CALL {
     WITH DISTINCT merged as merged
     RETURN COUNT(merged) AS total
 }
-CALL {
+CALL () {
     CALL db.index.fulltext.queryNodes("search_index", $query_string)
     YIELD node AS hit, score
     OPTIONAL MATCH (n:AdditiveThis|AdditiveThat|AdditiveOther|ExtractedThis|ExtractedThat|ExtractedOther)-[:stableTargetId]->(merged:MergedThis|MergedThat|MergedOther)
@@ -216,8 +216,8 @@ CALL {
         AND ANY(label IN labels(merged) WHERE label IN $labels)
     WITH DISTINCT merged as merged
     OPTIONAL MATCH (n)-[:stableTargetId]->(merged)
-    CALL {
-        WITH n
+    WITH n, merged
+    CALL (n) {
         OPTIONAL MATCH (n)-[r]->(referenced:MergedThis|MergedThat|MergedOther)
         RETURN CASE WHEN referenced IS NOT NULL THEN {
             label: type(r),
@@ -225,7 +225,6 @@ CALL {
             value: referenced.identifier
         } ELSE NULL END as ref
     UNION
-        WITH n
         OPTIONAL MATCH (n)-[r]->(nested:Link|Text|Location)
         RETURN CASE WHEN nested IS NOT NULL THEN {
             label: type(r),
@@ -246,21 +245,21 @@ RETURN collect(merged) AS items, total;""",
             False,
             False,
             """\
-CALL {
+CALL () {
     OPTIONAL MATCH (n:AdditiveThis|AdditiveThat|AdditiveOther|ExtractedThis|ExtractedThat|ExtractedOther)-[:stableTargetId]->(merged:MergedThis|MergedThat|MergedOther)
     WHERE
         ANY(label IN labels(merged) WHERE label IN $labels)
     WITH DISTINCT merged as merged
     RETURN COUNT(merged) AS total
 }
-CALL {
+CALL () {
     OPTIONAL MATCH (n:AdditiveThis|AdditiveThat|AdditiveOther|ExtractedThis|ExtractedThat|ExtractedOther)-[:stableTargetId]->(merged:MergedThis|MergedThat|MergedOther)
     WHERE
         ANY(label IN labels(merged) WHERE label IN $labels)
     WITH DISTINCT merged as merged
     OPTIONAL MATCH (n)-[:stableTargetId]->(merged)
-    CALL {
-        WITH n
+    WITH n, merged
+    CALL (n) {
         OPTIONAL MATCH (n)-[r]->(referenced:MergedThis|MergedThat|MergedOther)
         RETURN CASE WHEN referenced IS NOT NULL THEN {
             label: type(r),
@@ -268,7 +267,6 @@ CALL {
             value: referenced.identifier
         } ELSE NULL END as ref
     UNION
-        WITH n
         OPTIONAL MATCH (n)-[r]->(nested:Link|Text|Location)
         RETURN CASE WHEN nested IS NOT NULL THEN {
             label: type(r),
@@ -387,8 +385,8 @@ def test_fetch_identities(
         (
             ["personInCharge", "meetingScheduledBy", "agendaSignedOff"],
             """\
-MATCH (source:ExtractedThat {identifier: $identifier})-[stableTargetId:stableTargetId]->({identifier: $stable_target_id})
-CALL {
+MATCH (source:ExtractedThat {identifier: $identifier})-[:stableTargetId]->({identifier: $stable_target_id})
+CALL (source) {
     WITH source
     MATCH (target_0 {identifier: $ref_identifiers[0]})
     MERGE (source)-[edge:personInCharge {position: $ref_positions[0]}]->(target_0)
@@ -405,8 +403,7 @@ CALL {
     RETURN edge
 }
 WITH source, count(edge) as merged, collect(edge) as edges
-CALL {
-    WITH source, edges
+CALL (source, edges) {
     MATCH (source)-[outdated_edge]->(:MergedThis|MergedThat|MergedOther)
     WHERE NOT outdated_edge IN edges
     DELETE outdated_edge
@@ -417,13 +414,12 @@ RETURN merged, pruned, edges;""",
         (
             [],
             """\
-MATCH (source:ExtractedThat {identifier: $identifier})-[stableTargetId:stableTargetId]->({identifier: $stable_target_id})
-CALL {
+MATCH (source:ExtractedThat {identifier: $identifier})-[:stableTargetId]->({identifier: $stable_target_id})
+CALL (source) {
     RETURN null as edge
 }
 WITH source, count(edge) as merged, collect(edge) as edges
-CALL {
-    WITH source, edges
+CALL (source, edges) {
     MATCH (source)-[outdated_edge]->(:MergedThis|MergedThat|MergedOther)
     WHERE NOT outdated_edge IN edges
     DELETE outdated_edge
@@ -468,8 +464,7 @@ ON MATCH SET value_2 += $nested_values[2]
 WITH current,
     [edge_0, edge_1, edge_2] as edges,
     [value_0, value_1, value_2] as values
-CALL {
-    WITH current, values
+CALL (current, values) {
     MATCH (current)-[]->(outdated_node:Link|Text|Location)
     WHERE NOT outdated_node IN values
     DETACH DELETE outdated_node
@@ -488,8 +483,7 @@ ON MATCH SET current += $on_match
 WITH current,
     [] as edges,
     [] as values
-CALL {
-    WITH current, values
+CALL (current, values) {
     MATCH (current)-[]->(outdated_node:Link|Text|Location)
     WHERE NOT outdated_node IN values
     DETACH DELETE outdated_node