+++ CHANGELOG.md
@@ -19,6 +19,31 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security
 
+## [0.23.0] - 2024-11-19
+
+### Changes
+
+- improve error handling by returning validation issues for InconsistentGraphErrors
+- pin mex-release to 0.3.0
+
+## [0.22.0] - 2024-11-08
+
+### Changes
+
+- upgrade mex-common and mex-model dependencies to metadata model v3
+- apply additional linters in prep for `all` ruff linters
+- mute warnings about labels used in queries but missing in graph
+- split up search_merged_items_in_graph for better readability
+- update cypher queries to use `CALL` clauses with correct variable scope
+- BREAKING: drop support for neo4j server version 5.6 and lower
+
+## [0.21.0] - 2024-10-31
+
+### Changes
+
+- upgrade mex-common and mex-model dependencies to metadata model v3
+- apply additional linters in prep for `all` ruff linters
+
 ## [0.20.0] - 2024-10-18
 
 ### Changes